% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_modules.R
\name{assemble_modules}
\alias{assemble_modules}
\title{assemble_modules}
\usage{
assemble_modules(clustered_genes)
}
\arguments{
\item{clustered_genes}{Gene clustering results from blockclustering solution inside run_ICLite}
}
\value{
Gene modules
}
\description{
Creates modules from gene clustering results
}
