% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_to_mod_cor.R
\name{cell_to_mod_cor}
\alias{cell_to_mod_cor}
\title{cell_to_mod_cor}
\usage{
cell_to_mod_cor(score_mat)
}
\arguments{
\item{score_mat}{A score matrix generated by mod_score_matrix_assembly.}
}
\value{
A matrix of Spearman rho values
}
\description{
Calculates Spearman correlation between module scoring and cell log ratio.  Operates on "immune_cell_logratios" input to "run_ICLite"
}
