% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_binary_mat.R
\name{create_binary_mat}
\alias{create_binary_mat}
\title{create_binary_mat}
\usage{
create_binary_mat(rho_cutoff, min_connectivity)
}
\arguments{
\item{rho_cutoff}{A vector of rho exclusion values ranging from 0.1 to 0.9.}

\item{min_connectivity}{A vector of minimum connectivity values.  Higher cutoffs will result in smaller gene modules}
}
\value{
A binary matrix based off input cutoff parameters
}
\description{
Creates a binary matrix based off rho cutoff and connectivity values input into run_ICLite
}
