% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ICL.R
\name{extract_ICL}
\alias{extract_ICL}
\title{extract_ICL}
\usage{
extract_ICL(solution)
}
\arguments{
\item{solution}{Gene blockclustering solution from inside run_ICLite}
}
\value{
ICL value
}
\description{
Extracts ICL value from blockclustering solution inside run_ICLite
}
