% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cor_mat.R
\name{make_cor_mat}
\alias{make_cor_mat}
\title{make_cor_mat}
\usage{
make_cor_mat(solution)
}
\arguments{
\item{solution}{Gene blockclustering solution from inside run_ICLite}
}
\value{
Matrix of rho estimates from Spearman's rho calculation between gene module scoring and immune cell log ratios
}
\description{
Calculated matrix of rho values for correlation between gene modules and cells
}
