% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gene_cor_mat.R
\name{make_gene_cor_mat}
\alias{make_gene_cor_mat}
\title{make_gene_cor_mat}
\usage{
make_gene_cor_mat(expr_data)
}
\arguments{
\item{expr_data}{A matrix of normalized gene expression data where columns represent individuals and rows represent features (e.g. genes)}
}
\value{
A matrix of Spearman rho values
}
\description{
Contructs a symmetric correlation matrix from the input gene expression data
}
