% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_list.R
\name{make_input_list}
\alias{make_input_list}
\title{make_input_list}
\usage{
make_input_list(input_connectivities, input_rho)
}
\value{
List of input combinations
}
\description{
Contructs list of input parameters
}
