% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_size_penalty_calc.R
\name{matrix_size_penalty_calc}
\alias{matrix_size_penalty_calc}
\title{matrix_size_penalty_calc}
\usage{
matrix_size_penalty_calc(solution)
}
\arguments{
\item{solution}{Gene blockclustering solution from inside run_ICLite}
}
\value{
Unscaled penalty factor
}
\description{
Calculated matrix size penalty for optimal solution identification
}
