% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_score_matrix_assembly.R
\name{mod_score_matrix_assembly}
\alias{mod_score_matrix_assembly}
\title{mod_score_matrix_assembly}
\usage{
mod_score_matrix_assembly(mod_list)
}
\arguments{
\item{mod_list}{A list of character strings indicating gene module membership}
}
\value{
A matrix of module gene expression across individuals
}
\description{
Wrapper for scoring module expression across individuals
}
