% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_successful_connections.R
\name{quantify_successful_connections}
\alias{quantify_successful_connections}
\title{quantify_successful_connections}
\usage{
quantify_successful_connections(pval_mat, cor_mat)
}
\arguments{
\item{pval_mat}{p-values for gene to cell interations}

\item{cor_mat}{rho values for gene to cell interations}
}
\value{
Integer value of positive correlations
}
\description{
Tallies significant positive correlations
}
