% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ICLite.R
\name{run_ICLite}
\alias{run_ICLite}
\title{run_ICLite: Wrapper function for ICLite package}
\usage{
run_ICLite(
  gene_expression_data,
  immune_cell_logratios,
  input_connectivities,
  input_rho,
  number_of_clusters
)
}
\arguments{
\item{gene_expression_data}{A matrix of normalized gene expression data where columns represent individuals and rows represent features (e.g. genes)}

\item{immune_cell_logratios}{A matrix of corresponding cell log ratios where rows represent individuals and columns represent features (e.g. cells).  Transformation of percentage values using the "compositions" package is recommended prior to use}

\item{input_connectivities}{A vector of minimum connectivity values.  Higher cutoffs will result in smaller gene modules}

\item{input_rho}{A vector of rho exclusion values ranging from 0.3 to 0.9.  Gene correlations below this value are converted to 0 in binary space while those above are converted to 1.  Higher cutoffs will result in smaller gene modules}

\item{number_of_clusters}{A vector of assumed number of clusters to be used for blockclustering}
}
\value{
Module to gene connectivity solution and gene module memberships to working directory and global environment
}
\description{
Runs the ICLite algorithm based on input parameters provided
by user.  Creates output graphical files based on gene module and cell interactions,
CSV files of gene modules and global environment object "gene_module_lists" that
may be used for downstream Gene Ontology analysis.  Example data can be loaded using
load_IMSA_data()

Finding the optimal solution for a data set may require running multiple iterations
of ICLite with varying input parameters.  Running separate tests for correlation
between genes in a data set is recommended to determine minimum values for input
rho exclusion.  For general practice, we recommend using no lower than 0.4.  The
number of assumed gene clusters should be considered in relation to the total size
of the transcriptional data set.  Though ICLite does penalize for over-clustering,
it will only consider solutions from the input vector.  Therefore, initial runs
may benefit from a broad array of values that may be narrowed on successive iterations.

Users should include a vector of at least 2 input parameters for minimum connectivity,
rho cutoff and number of clusters.
}
