% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solution_by_num_clust.R
\name{solution_by_num_clust}
\alias{solution_by_num_clust}
\title{solution_by_num_clust}
\usage{
solution_by_num_clust(number_of_clusters)
}
\arguments{
\item{number_of_clusters}{Input vector for number of clusters}
}
\value{
A list of clustering solutions
}
\description{
Runs blockclustering for binary matrices created inside run_ICLite using varied numbers of clusters
}
