% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_cell_connections.R
\name{test_for_cell_connections}
\alias{test_for_cell_connections}
\title{test_for_cell_connections}
\usage{
test_for_cell_connections(solution)
}
\arguments{
\item{solution}{A blockclustering solution}
}
\value{
An integer value of total positive correlations between cells and gene modules
}
\description{
Quantifies number of positive correlations between gene modules and cells for an input gene clustering solution
}
