% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_IMSA_data.R
\name{load_IMSA_data}
\alias{load_IMSA_data}
\title{load_IMSA_data}
\usage{
load_IMSA_data()
}
\value{
loads immune_cell_logratios and gene_expression_data objects into global environment
}
\description{
Loads gene expression and immune log ratios from the IMSA data set
}
\examples{

#' run_ICLite(gene_expression_data, immune_cell_logratios, input_connectivities,
           input_rho, number_of_clusters)

}
