% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_input_list.R
\name{make_input_list}
\alias{make_input_list}
\title{make_input_list}
\usage{
make_input_list(input_connectivities, input_rho)
}
\arguments{
\item{input_connectivities}{vector of input connectivity requirements.  Larger numbers create smaller modules}

\item{input_rho}{Vector of input rho exclusion values ranging from 0.1 to 0.9.  Anything below this cutoff is converted to a 0 on binary scale.  All interactions above are entered as 1 in the binary matrix.  Larger numbers create smaller modules.}
}
\value{
List of input combinations
}
\description{
Contructs list of input parameters
}
